DRY up duplicate code
