+++ app/views/ministries/_file_form.html.arb
@@ -1,11 +1,32 @@
 # vim:ft=ruby:
 
-form_for :spreadsheet, url: {action: action}, html: {multipart: true} do |f|
-  table do
-    tr do
-      td { label(for: :spreadsheet_file) { 'Select a File' } }
-      td { f.file_field :file }
+form_for "spreadsheet_#{action}", url: {action: action}, html: {multipart: true} do |f|
+  fieldset class: 'inputs' do
+    ol do
+      li do
+        f.label :file, 'Select a File'
+        f.file_field :file
+        para class: 'inline-hints' do
+          'Must be an .ods, .csv, .xls, or .xlsx file.'
+        end
+      end
+      li class: 'boolean input' do
+        label do
+          f.check_box :skip_first
+          span 'Skip Header Row'
+        end
+      end
     end
-    tr { td; td(style: 'text-align: right') { submit_tag 'Go!' } }
+  end
+
+  fieldset class: 'actions' do
+    ol do
+      li class: 'action input_action' do
+        submit_tag 'Submit'
+      end
+      li class: 'cancel' do
+        link_to 'Cancel', ministries_path
+      end
+    end
   end
 end
